Use bigint for Telegram identifiers

+++ backend/app/models/scan.py
@@ -1,6 +1,6 @@
 from datetime import datetime
 
-from sqlalchemy import DateTime, Integer, String, Text, func
+from sqlalchemy import BigInteger, DateTime, Integer, String, Text, func
 from sqlalchemy.orm import Mapped, mapped_column
 
 from app.core.database import Base
@@ -16,9 +16,9 @@ class ScanRecord(Base):
     level: Mapped[str] = mapped_column(String(16), index=True)
     threat_type: Mapped[str] = mapped_column(String(32))
     confidence: Mapped[int] = mapped_column(Integer, default=0)
-    telegram_chat_id: Mapped[int | None] = mapped_column(Integer, nullable=True, index=True)
-    telegram_user_id: Mapped[int | None] = mapped_column(Integer, nullable=True, index=True)
-    telegram_message_id: Mapped[int | None] = mapped_column(Integer, nullable=True, unique=True, index=True)
+    telegram_chat_id: Mapped[int | None] = mapped_column(BigInteger, nullable=True, index=True)
+    telegram_user_id: Mapped[int | None] = mapped_column(BigInteger, nullable=True, index=True)
+    telegram_message_id: Mapped[int | None] = mapped_column(BigInteger, nullable=True, unique=True, index=True)
     media_type: Mapped[str | None] = mapped_column(String(32), nullable=True)
     indicators_json: Mapped[str] = mapped_column(Text, default="[]")
     threat_intelligence_json: Mapped[str] = mapped_column(Text, default="[]")
